+++ ConsoleBalatro/Engine/Events/Args/EngineHandPlayDoneArgs.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -12,8 +13,8 @@ namespace ConsoleBalatro.Engine.Events.Args
         public PlayedHandType HandTypeThatWasPlayed { get; set; }
         public List<Card> CardsInPlayedHand { get; set; }
         public List<Card> CardsHeldInHand { get; set; }
-        public long CurrentTotalChips { get; set; }
-        public long RequiredChipsForBlind { get; set; }
+        public BigInteger CurrentTotalChips { get; set; }
+        public BigInteger RequiredChipsForBlind { get; set; }
         public bool PreventGameOverAndWinBlind { get; set; } = false;
 
     }
