+++ ConsoleBalatro/Engine/Events/Args/EngineSettingBaseHandScoreArgs.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,7 +9,7 @@ namespace ConsoleBalatro.Engine.Events.Args
 {
     public class EngineSettingBaseHandScoreArgs : EngineEventArgs
     {
-        public int BaseChipAmount { get; set; }
+        public BigInteger BaseChipAmount { get; set; }
         public double BaseMultAmount { get; set; }
     }
 }
